Avoid overflow in computation of whether a case statement is sparse.
If the difference between the minimum and maximum values in a case statement was greater than 32767, then the subtraction in this computation would overflow. That would cause the case statement to be treated as non-sparse, so a jump table would be generated for it, but that jump table would be very large, so it would overflow the object buffer and cause an error.